CMA-(CreationOfProjectStructureAndClasses) - (Project Structure Established & Preliminary Classes Created) - (Commit No.1)
I established a preliminary project structure by declaring the relevant packages(controller,dto,model,repository,service).

In each of those packages I declared preliminary version of classes that are going to be used throughout the application.

What is more I added project Lombok dependency inside Maven, so that I can make use of getters and setters if there is such a need.

Please note that this is just a preliminary version of the application and more what we would call a "prototype". The features will need to be corrected adequately.